perbaikan pemanggilan api
