+++ src/admin/components/TabProfile.jsx
@@ -3,11 +3,6 @@ import { Tabs } from 'antd'
 import VisionMisionForm from './VisionMisionForm'
 import KenapaSumod from './KenapaSumod'
 
-const onChange = ( key ) => {
-  console.log( key )
-}
-
-
 
 export default function TabProfile( { products, renderFormattedDescription, setSelectedProduct, setNewData } ) {
 
@@ -26,15 +21,15 @@ export default function TabProfile( { products, renderFormattedDescription, setS
       children: VisionMisionContent,
     },
     {
-      key: '3',
+      key: '2',
       label: 'Kenapa Sumod',
-      children: KenapaSumodContent // Insert KenapaSumod component here
+      children: KenapaSumodContent
     },
   ]
 
   return (
     <div className='py-5 mb-10'>
-      <Tabs defaultActiveKey="1" items={items} onChange={onChange} />
+      <Tabs defaultActiveKey="1" items={items} />
     </div>
   )
 }
